`fetch()`'s `input` argument is not optional
It's not optional in the browser DOM signature which this is emulating,
and it's not clear what the behavior of a `fetch()` should be if it were
given `undefined`. Type-wise, making this optional requires any client
code providing a `fetch()` function to handle the `undefined` value,
even though it's nonsensical.